Let there be code!
Initial commit

And so it was on this day, that the branch Students of Robot_Framework
was born into this world. It is today in which Students hath spoken,
"Hello World" and graced us with its presense. Oh Students! How will you
grace the world? Will you create code to last the test of time?